requirements.txt
django==3.2
django-environ==0.11.2
black==24.8.0
djangorestframework==3.15.2